user-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
python-jose[cryptography]==3.3.0
pydantic==2.6.4